check if file exists before copying to avoid exception

+++ src/main/java/de/tum/in/www1/artemis/service/programming/ProgrammingExerciseImportFromFileService.java
@@ -19,10 +19,7 @@ import org.springframework.web.multipart.MultipartFile;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import de.tum.in.www1.artemis.domain.Course;
-import de.tum.in.www1.artemis.domain.ProgrammingExercise;
-import de.tum.in.www1.artemis.domain.Repository;
-import de.tum.in.www1.artemis.domain.VcsRepositoryUrl;
+import de.tum.in.www1.artemis.domain.*;
 import de.tum.in.www1.artemis.domain.enumeration.RepositoryType;
 import de.tum.in.www1.artemis.service.*;
 import de.tum.in.www1.artemis.service.connectors.GitService;
@@ -94,7 +91,7 @@ public class ProgrammingExerciseImportFromFileService {
      *
      * @param importExerciseDir the directory where the extracted zip file is located
      **/
-    private void copyEmbeddedFiles(Path importExerciseDir) {
+    private void copyEmbeddedFiles(Path importExerciseDir) throws IOException {
         Path embeddedFilesDir = importExerciseDir.resolve("files");
 
         if (!Files.exists(embeddedFilesDir)) {
@@ -103,11 +100,11 @@ public class ProgrammingExerciseImportFromFileService {
         try (var embeddedFiles = Files.list(embeddedFilesDir)) {
             for (var file : embeddedFiles.toList()) {
                 var targetPath = Path.of(FilePathService.getMarkdownFilePath(), file.getFileName().toString());
-                Files.copy(file, targetPath);
+                // we need this check because the detection if a file exists of Files.copy seems not to work properly
+                if (!Files.exists(targetPath)) {
+                    Files.copy(file, targetPath);
+                }
             }
-        }
-        catch (IOException e) {
-            throw new RuntimeException(e);
         }
     }
